Refactor import paths for AuthController and DocumentController; add DocumentController implementation for document operations
